DIABLO-1311, /home accessibility: manageable render logic for eligible and ineligible courses
